Create user index page

+++ src/th/ac/kmitl/it/foodbook/servlets/users/DeleteServlet.java
@@ -9,9 +9,15 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import javax.sql.DataSource;
 
+import th.ac.kmitl.it.foodbook.beans.RecipeCategory;
+import th.ac.kmitl.it.foodbook.daos.RecipeCategoriesDAO;
+import th.ac.kmitl.it.foodbook.daos.RecipesDAO;
 import th.ac.kmitl.it.foodbook.daos.UsersDAO;
+import th.ac.kmitl.it.foodbook.utils.Alert;
+import th.ac.kmitl.it.foodbook.utils.Alert.AlertTypes;
 
 @WebServlet("/users/delete")
 public class DeleteServlet extends HttpServlet {
@@ -23,21 +29,41 @@ public class DeleteServlet extends HttpServlet {
     }
     
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        String userIdString = request.getParameter("user_id");
-        long userId = Long.parseLong(userIdString);
+        String[] userIdsString = request.getParameterValues("user_id");
+        
+        boolean isSuccess = false;
         
         DataSource ds = (DataSource) request.getServletContext().getAttribute("ds");
+        
+        HttpSession session = request.getSession();
         
         try {
             Connection conn = ds.getConnection();
             
-            UsersDAO usersDAO = new UsersDAO(conn);
-            usersDAO.delete(userId);
+            if (userIdsString == null) {
+                response.sendRedirect("/users/index");
+                return;
+            }
+            
+            for (String userIdString : userIdsString) {
+                long userId = Long.parseLong(userIdString);
+                
+                UsersDAO usersDAO = new UsersDAO(conn);
+                isSuccess = usersDAO.delete(userId);
+            }
             
             conn.close();
         } catch (SQLException e) {
             e.printStackTrace();
             response.sendError(500);
+        }
+        
+        if (isSuccess) {
+            session.setAttribute("alert", new Alert(AlertTypes.SUCCESS, "Deleted Successfully!"));
+            response.sendRedirect("/users/index");
+        } else {
+            session.setAttribute("alert", new Alert(AlertTypes.DANGER, "Deleted Unsuccessfully!"));
+            response.sendRedirect("/users/index");
         }
     }
